# Optimize `vec::retain` performance 