@the8472 2021-12-04
retain is used in some places in the compiler, some are fairly warm code, let's see if it has an impact.

@bors try @rust-timer queue