@kageru 2021-12-04
Testing this PR on the system from my report in #91497 (Ryzen 5900X with Linux).
Old:
```
test vec::bench_retain_100000                            ... bench:     188,266 ns/iter (+/- 4,000)
test vec::bench_retain_whole_100000                      ... bench:      42,654 ns/iter (+/- 1,221)
```
New:
```
test vec::bench_retain_100000                            ... bench:      41,626 ns/iter (+/- 148)
test vec::bench_retain_iter_100000                       ... bench:      41,319 ns/iter (+/- 1,042)
test vec::bench_retain_whole_100000                      ... bench:      21,833 ns/iter (+/- 164)
```
Looking much better.