@TennyZhuang 2021-12-04
Interesting, I don't even know why LLVM can't optimize this code, but the benchmark is very convincing.

LGTM